chore : layout.tsx를 활용한 레이아웃 분리 실패
